experiments/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "experiments"                   # 发行名（pip里显示的名字）——你坚持用就可以
version = "0.1.0"
description = "Cirq experiments: scripts & configs"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6",
  # 按需加上 numpy/pandas 等
  # cirq 可不写（你本地以 editable 安装 dev/Cirq），
  # 或写下限要求比如 "cirq>=1.3"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# 把 YAML 一起打包（wheel/sdist 都包含）
[tool.setuptools.package-data]
"experiments" = ["configs/*.yaml", "configs/*.yml", "configs/*.json"]

# 可选：命令行入口
[project.scripts]
run-exp = "experiments.scripts.run_exp1_experiment:main"
save-results = "experiments.scripts.save_results:main"
